PORTALS-3583
Lazy-load routes for Challenge Portal
