+++ apps/portals/challenges/src/pages/ChallengeDetailsPage.tsx
@@ -1,0 +1,46 @@
+import { challengeDetailsPageTabConfig } from '@/config/challengeDetailsPageTabConfig'
+import { challengeProjectsSql } from '@/config/resources'
+import { challengeCardConfiguration } from '@/config/synapseConfig/challenges'
+import ChallengeDetailPageWrapper from '@sage-bionetworks/synapse-portal-framework/components/challenges/ChallengeDetailPageWrapper'
+import { DetailsPageTabs } from '@sage-bionetworks/synapse-portal-framework/components/DetailsPage/DetailsPageTabs'
+import DetailsPage from '@sage-bionetworks/synapse-portal-framework/components/DetailsPage/index'
+import { useGetPortalComponentSearchParams } from '@sage-bionetworks/synapse-portal-framework/utils/UseGetPortalComponentSearchParams'
+import { ColumnSingleValueFilterOperator } from '@sage-bionetworks/synapse-types'
+import { Outlet, RouteObject } from 'react-router'
+import {
+  CardContainerLogic,
+  ErrorPage,
+  SynapseErrorType,
+} from 'synapse-react-client'
+
+export const challengeDetailsPageTabRoutes: RouteObject[] = []
+
+function ChallengeDetailsPage() {
+  const { id } = useGetPortalComponentSearchParams()
+
+  if (!id) {
+    return <ErrorPage type={SynapseErrorType.NOT_FOUND} gotoPlace={() => {}} />
+  }
+  return (
+    <>
+      <div className={'challengeDetailPageHeaderCard'}>
+        <CardContainerLogic
+          cardConfiguration={{ ...challengeCardConfiguration, isHeader: true }}
+          sql={challengeProjectsSql}
+          searchParams={{ id }}
+        />
+      </div>
+      {/*Register For Challenge component is here, and the wrapper positions it properly in the header card*/}
+      <ChallengeDetailPageWrapper />
+      <DetailsPage
+        sql={challengeProjectsSql}
+        sqlOperator={ColumnSingleValueFilterOperator.EQUAL}
+      >
+        <DetailsPageTabs tabConfig={challengeDetailsPageTabConfig} />
+        <Outlet />
+      </DetailsPage>
+    </>
+  )
+}
+
+export default ChallengeDetailsPage
